requirements.txt
langchain-openai
python-dotenv
langchain_community
faiss-cpu
langchain_text_splitters
